+++ src/TitleBindPieces/index.js
@@ -1,0 +1,1 @@
+export { default as TitleBindPiecesContainer } from './TitleBindPiecesContainer';